+++ web/sw.js
@@ -1,4 +1,4 @@
-const BUILD_ID = "air-drow-v760-final-clean-delivery";
+const BUILD_ID = "air-drow-v761-interaction-vision-tutorial";
 const CACHE_PREFIX = "air-drow-runtime-";
 const STATIC_CACHE = `${CACHE_PREFIX}${BUILD_ID}`;